actually parsing freq

+++ stm_workspace/hydrophone/Core/Inc/ping.h
@@ -31,6 +31,6 @@
 uint8_t has_ping(uint32_t*, uint32_t, uint32_t);
 uint32_t get_power_at_target_frequency(uint32_t* buff, uint32_t size);
 uint32_t get_total_power(uint32_t* buff, uint32_t size);
-uint32_t get_frequency(uint32_t* buff, uint32_t size, float32_t fs);
+uint32_t get_frequency(float32_t* buff, uint32_t size, float32_t fs);
 
 #endif /* INC_PING_H_ */
